Move read/write code to separate modules
