+++ lasio/defaults.py
@@ -1,4 +1,3 @@
-import re
 
 import numpy as np
 
@@ -8,14 +7,6 @@ from las_items import (
 
 
 
-URL_REGEXP = re.compile(
-    r'^(?:http|ftp)s?://'  # http:// or https://
-    r'(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}'
-    r'\.?|[A-Z0-9-]{2,}\.?)|'  # (cont.) domain...
-    r'localhost|'  # localhost...
-    r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})'  # ...or ip
-    r'(?::\d+)?'  # optional port
-    r'(?:/?|[/?]\S+)$', re.IGNORECASE)
 
 
 DEFAULT_ITEMS = {
